Update branch "sponzori traka"
I updated this branch and it wont affect main fls.ba page, this is just for Haris to see moveable sponsors and if we are going to proceed with implementing it.
If not we can revert back to just <Partners /> and leave that on our main page which is what I will commit on main branch.